feat(issue): add subcommand to view issue details and comments
This command is designed to display detailed information about a specific issue
in Arch Linux packaging projects. It gathers and pretty prints all relevant
data about the issue, providing a comprehensive view that includes the issue's
description, status as well as labels and creation date.

By default, the command operates within the current directory, but users have
the option to specify a different package base. Additionally, users can choose
to view the issue in a web browser for a more interactive experience.

Component: pkgctl issue view
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>