feat(issue): add subcommand to close issues
This command is used to close an issue in Arch Linux packaging projects. It
finalizes the issue by marking it as resolved and optionally providing a reason
for its closure.

By default, the command operates within the current directory, but users have
the option to specify a different package base.

Users can provide a message directly through the command line to explain the
reason for closing the issue. Additionally, a specific resolution label can be
set to categorize the closure reason, with the default label being "completed."

Component: pkgctl issue close
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>