feat(issue): add subcommand to move issues between projects
The move command allows users to transfer an issue from one project to another
within the Arch Linux packaging group. This is useful when an issue is
identified to be more relevant or better handled in a different project.

By default, the command operates within the current directory, but users can
specify a different package base from which to move the issue.

Users must specify the issue ID (IID) and the destination package to which the
issue should be moved. A comment message explaining the reason for the move can
be provided directly through the command line.

Component: pkgctl issue move
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>