feat(version): graceful config for packages without remote sources
Setup a blank config for packages without remote sources. This is
helpful so other commands like `pkgctl version check` operate gracefully
as well as we have easy way to find packages that miss nvchecker
config.

This must only be used for cases without an upstream, please reach out
to the developer team for guidance regarding upstream sources that are
hard to configure.

Component: pkgctl version setup
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>