feat(build): Invoke pkgctl offload server subcommand over ssh
Use a client-server script invocation instead of piping complex shell
commands as pure strings to the remote. Now we call a server use only
component on the remote server to execute the desired build.

This change allows for properly developing the remote command without
golfing around shell strings and while having appropriate completion,
syntax highlight and LSP support. As a side-effect we are now not
depending on a POSIX compliant login shell anymore, as we simply invoke
a pkgctl subcommand schema.

Component: pkgctl build
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>